refactor: move ffmpeg/ into cobalt_converter package
Update all internal imports to use cobalt_converter.ffmpeg.* prefix
