+++ cobalt_converter/ffmpeg/__init__.py
@@ -1,0 +1,3 @@
+from cobalt_converter.ffmpeg.resolver import FFmpegResolver
+
+__all__ = ["FFmpegResolver"]
