--- ffmpeg/__init__.py
@@ -1,3 +1,0 @@
-from ffmpeg.resolver import FFmpegResolver
-
-__all__ = ["FFmpegResolver"]
